requirements.txt
datasets>=3.5.0
transformers>=4.51.0
openai>=1.72.0
matplotlib>=3.7.0
seaborn>=0.13.0
numpy>=1.23.0
pandas>=2.0.0
tqdm>=4.66.0
pathlib>=1.0.1
python-dotenv>=1.0.0
argparse>=1.4.0
pytest>=7.4.0
ruff>=0.1.0
streamlit>=1.38.0
